Update JS_Prac1.html
Write JavaScript code to find sum of series
1/2+3/4+5/6+7/8+…..
Take input for number of terms in series in an input textbox. (e.g. - number of terms in the example above is 4). Also write validation script to check whether the input field is not blank and that only numeric values are entered.